komis_excel_updater/requirements.txt
openpyxl>=3.1.5
requests>=2.32.3
pandas>=2.2.2
beautifulsoup4>=4.12.3
lxml>=5.2.2
python-dateutil>=2.9.0.post0
playwright>=1.52.0
